CNN/BovinaBreedPrediction/app-aws/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow-cpu==2.20.0
python-multipart
boto3
mangum

